@@ -46,7 +46,8 @@
     "babel-istanbul": "^0.12.2",
     "babel-loader": "^6.4.1",
     "babel-plugin-react-display-name": "^2.0.0",
-    "babel-preset-es2015": "^6.22.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-2": "^6.22.0",
@@ -66,7 +67,6 @@
     "eslint-watch": "^3.0.1",
     "eventsource-polyfill": "0.9.6",
     "expect": "1.19.0",
-    "express": "4.13.4",
     "extract-text-webpack-plugin": "1.0.1",
     "faker": "^4.1.0",
     "file-loader": "0.8.5",
@@ -78,6 +78,7 @@
     "nock": "8.0.0",
     "npm-run-all": "1.8.0",
     "open": "0.0.5",
+    "prop-types": "^15.6.2",
     "react-addons-test-utils": "^15.0.2",
     "redux-immutable-state-invariant": "1.2.3",
     "redux-mock-store": "1.0.2",
@@ -97,8 +98,6 @@
     "babel-polyfill": "6.8.0",
     "babel-register": "^6.23.0",
     "bcrypt-nodejs": "0.0.3",
-    "body-parser": "^1.17.1",
-    "bootstrap": "3.3.6",
     "cheerio": "^0.22.0",
     "classnames": "^2.2.5",
     "colors": "^1.1.2",
@@ -117,20 +116,15 @@
     "nightwatch": "^0.9.14",
     "pg": "^6.1.4",
     "pg-hstore": "^2.3.2",
-    "react": "15.0.2",
-    "react-dom": "15.0.2",
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2",
     "react-froala-wysiwyg": "^2.5.1",
-    "react-redux": "4.4.5",
-    "react-redux-sweetalert": "^1.0.0",
-    "react-router": "2.4.0",
-    "react-router-dom": "^4.0.0",
-    "react-router-redux": "4.0.4",
-    "react-rte": "^0.11.0",
-    "react-select": "^1.0.0-rc.3",
-    "redux": "^3.6.0",
-    "redux-devtools-extension": "^2.13.0",
-    "redux-router": "^2.1.2",
-    "redux-thunk": "^2.2.0",
+    "react-redux": "^5.1.1",
+    "react-redux-sweetalert": "^1.0.2",
+    "react-router-dom": "^4.3.1",
+    "redux": "^4.0.1",
+    "redux-devtools-extension": "^2.13.5",
+    "redux-thunk": "^2.3.0",
     "sequelize": "^3.30.2",
     "sequelize-cli": "^2.5.1",
     "shortid": "^2.2.8",
