@@ -1,10 +1,6 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable import/no-extraneous-dependencies */
-// More info on Webpack's Node API here: https://webpack.github.io/docs/node.js-api.html
-// Allowing console calls below since this is a build file.
 /* eslint-disable no-console */
 import webpack from 'webpack';
-import colors from 'colors';
+import 'colors';
 
 import webpackConfig from '../webpack.config.prod';
 
