@@ -1,9 +1,6 @@
-// This script copies src/index.html into /dist/index.html
-// This is a good example of using Node and cheerio to do a simple file transformation.
-// In this case, the transformation is useful since we only use a separate css file in prod.
 import fs from 'fs';
 import cheerio from 'cheerio';
-import colors from 'colors';
+import 'colors';
 
 /* eslint-disable no-console */
 
@@ -17,7 +14,7 @@ fs.readFile('client/index.html', 'utf8', (err, markup) => {
   // since a separate spreadsheet is only utilized for the production build, need to dynamically add this here.
   $('head').prepend('<link rel="stylesheet" href="styles.css">');
 
-  fs.writeFile('dist/index.html', $.html(), 'utf8', function (err) {
+  fs.writeFile('dist/index.html', $.html(), 'utf8', (err) => {
     if (err) {
       return console.log(err);
     }
