@@ -1,7 +1,4 @@
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint-disable no-unused-vars */
-
-import colors from 'colors';
+import 'colors';
 
 /* eslint-disable no-console */
 
